refactor: move service types to a services folder
