+++ .github/cicd/core/services/DeleteAPIVersionService.ts
@@ -1,5 +1,5 @@
 import { Directory } from "io/Directory.ts";
-import { Guard } from "./Guard.ts";
+import { Guard } from "../Guard.ts";
 import { VersionSideBarService } from "./VersionSidebarService.ts";
 import { VersionsFileService } from "./VersionsFileService.ts";
