+++ .github/cicd/core/services/MarkdownFileContentService.ts
@@ -4,7 +4,7 @@ import { Path } from "io/Path.ts";
 import { HTMLService } from "./HTMLService.ts";
 import { MarkdownHeaderService } from "./MarkdownHeaderService.ts";
 import { MarkdownService } from "./MarkdownService.ts";
-import { Utils } from "./Utils.ts";
+import { Utils } from "../Utils.ts";
 
 export class MarkdownFileContentService {
 	private readonly markDownService: MarkdownService;
