+++ .github/cicd/core/services/MarkdownHeaderService.ts
@@ -1,5 +1,5 @@
-import { Guard } from "./Guard.ts";
-import { Utils } from "./Utils.ts";
+import { Guard } from "../Guard.ts";
+import { Utils } from "../Utils.ts";
 
 /**
  * Processes markdown headers.
