+++ .github/cicd/scripts/delete-oldest-api-docs.ts
@@ -1,7 +1,7 @@
-import { DeleteAPIVersionService } from "../core/DeleteAPIVersionService.ts";
 import chalk from "../deps.ts";
-import { VersionsFileService } from "../core/VersionsFileService.ts";
 import { Utils } from "../core/Utils.ts";
+import { DeleteAPIVersionService } from "../core/services/DeleteAPIVersionService.ts";
+import { VersionsFileService } from "../core/services/VersionsFileService.ts";
 
 /**
  * DESCRIPTION: This script is used as part of the api docs release cicd process and is used
