+++ .github/cicd/scripts/update-website-version.ts
@@ -1,4 +1,4 @@
-import { UpdateWebsiteVersionService } from "../core/UpdateWebsiteVersionService.ts";
+import { UpdateWebsiteVersionService } from "../core/services/UpdateWebsiteVersionService.ts";
 
 const service = new UpdateWebsiteVersionService();
 service.updateVersion();
